node creation implemented; perftab somehow working; fixed problem with overview tab; upgraded to simplexml 2.4.1; mssql and odbc db drivers updated; other minor changes

+++ src/java/netxms-client/src/main/java/org/netxms/client/datacollection/GraphSettings.java
@@ -864,11 +864,11 @@ public class GraphSettings
 		switch(timeUnit)
 		{
 			case TIME_UNIT_MINUTE:
-				return timeFrame * 60 * 1000;
+				return (long)timeFrame * 60L * 1000L;
 			case TIME_UNIT_HOUR:
-				return timeFrame * 60 * 60 * 1000;
+				return (long)timeFrame * 60L * 60L * 1000L;
 			case TIME_UNIT_DAY:
-				return timeFrame * 24 * 60 * 60 * 1000;
+				return (long)timeFrame * 24L * 60L * 60L * 1000L;
 		}
 		return 0;
 	}
